+++ src/main/java/com/hibernate/ferreteria/repositorio/Repo_articulos.java
@@ -4,6 +4,6 @@ import com.hibernate.ferreteria.entity.Articulo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 //JPARepository, para acceso a datos. Objeto Articulo, tipo a usar Long
-public interface Repo_articulos extends JpaRepository<Articulo, Long> {
+public interface Repo_articulos extends JpaRepository<Articulo, Integer> {
 
 }
